+++ pkg/vm/engine/tae/txn/txnimpl/node.go
@@ -152,5 +152,5 @@ func (n *node) GetColumnDataById(idx int) (view *model.ColumnView, err error) {
 
 func (n *node) Prefetch(idxes []uint16) error {
 	key := n.meta.GetMetaLoc()
-	return blockio.Prefetch(idxes, []uint32{key.ID()}, n.fs.Service, key)
+	return blockio.Prefetch(idxes, []uint16{key.ID()}, n.fs.Service, key)
 }
